Time: 2 ms (97.85%), Space: 57.2 MB (87.1%) - LeetHub

+++ 3152-maximum-value-of-an-ordered-triplet-ii/3152-maximum-value-of-an-ordered-triplet-ii.java
@@ -6,7 +6,7 @@ class Solution {
         int ijMax = 0; // Max value of (nums[i] - nums[j])
 
         for (int num : nums) {
-            answer = Math.max(ijMax * num, answer);
+            answer = Math.max((long) ijMax * num, answer);
 
             ijMax = Math.max(iMax - num, ijMax);
